+++ droit/plugins/output/math/main.py
@@ -1,5 +1,5 @@
 def calc(data, rpack):
 	equation = data[0].lower().replace("plus", "+").replace("minus", "-")
 	equation = equation.replace("^", "**").replace("x", "*").replace(",", ".")
-	return eval(equation)
+	return eval(equation), rpack